ai-service/requirements.txt
fastapi>=0.109.1
uvicorn==0.27.0
scikit-surprise==1.1.3
scikit-learn==1.4.0
pandas==2.2.0
numpy>=1.26.0,<2.0.0
joblib==1.3.0
python-dotenv==1.0.0
